Add files via upload
Atomic Memory Model

Abstract. The task to track down and remove memory leaks is known as one of the most tedious and unpleasant ones that a software engineer might encounter. In some cases once part of a program memory leaks it might become impossible to remove. Buffer overruns are another problem often even more difficult and unpleasant to solve. This paper suggests an effective way to eradicate the above issues and other problems related to memory in a computer program. This is achieved by enforcing a preemptive methodology for memory handling in addition to abstraction of the memory and representing it as encapsulated entity with three properties: class, semantics and origin. The Atomic Memory Model is a consistent and elegant methodology as opposed to garbage collecting. It significantly improves the quality of code making most of the issues related to memory intrinsically impossible to occur and increasing the speed and ease of development.

For detailed explanation of the Atomic Memory Model please read the paper published in the Information Technologies and Control Magazine issue 02-2009: http://www.acad.bg/rismim/itc/sub/archiv/Paper1_2_2009.pdf

To use the Atomic Memory Model

1. Download the source code
2. Include the necessary header file,e.g.

#include "StringEx.h"
or
#include "MAtom.h"